requirements.txt
openai
python-dotenv
fastapi
uvicorn[standard]
pydantic
pytest
pytest-asyncio
httpx
pytest-mock
sqlalchemy
pymysql
cryptography
numpy
scikit-learn
python-multipart
